+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.15
 -----------------
 - Security: update PHPMailer to version 2.9.21
+- Fix #176 SQL error More than 1 row returned by a subquery in Rollover.php
 
 Changes in 2.9.14
 -----------------